+++ src/server/MultimediaDet/MultimediaDetCustom.ts
@@ -1,18 +1,11 @@
-import { ApiModule } from "../ApiModule";
-import { iNode } from "../iNode";
-import { MultimediaDet } from "../../crosscommon/entities/MultimediaDet";
 import { ApiServer } from "../ApiServer";
+import { MultimediaDet } from "../../crosscommon/entities/MultimediaDet";
 
 export class MultimediaDetCustom {
   private api: ApiServer = new ApiServer(new MultimediaDet());
 
-  list = (node: iNode) => {
-    let api: ApiModule = new ApiModule(new MultimediaDet());
-
-    api.list({ q: node.request.query["q"] }).then(response => {
-      node.response.end(JSON.stringify(response));
-    });
-  };
+  listRequestHandler = this.api.listRequestHandler;
+  list = this.api.list;
 
   createRequestHandler = this.api.createRequestHandler;
   create = this.api.create;