+++ src/server/MultimediaView/MultimediaViewCustom.ts
@@ -1,18 +1,11 @@
-import { ApiModule } from "../ApiModule";
-import { iNode } from "../iNode";
-import { MultimediaView } from "../../crosscommon/entities/MultimediaView";
 import { ApiServer } from "../ApiServer";
+import { MultimediaView } from "../../crosscommon/entities/MultimediaView";
 
 export class MultimediaViewCustom {
   private api: ApiServer = new ApiServer(new MultimediaView());
 
-  list = (node: iNode) => {
-    let api: ApiModule = new ApiModule(new MultimediaView());
-
-    api.list({ q: node.request.query["q"] }).then(response => {
-      node.response.end(JSON.stringify(response));
-    });
-  };
+  listRequestHandler = this.api.listRequestHandler;
+  list = this.api.list;
 
   createRequestHandler = this.api.createRequestHandler;
   create = this.api.create;